+++ src/main/java/org/folio/entlinks/integration/kafka/AuthorityEventListener.java
@@ -1,8 +1,14 @@
 package org.folio.entlinks.integration.kafka;
 
+import static java.util.Collections.singletonList;
+import static org.apache.commons.lang3.StringUtils.isBlank;
+import static org.folio.entlinks.utils.HeaderUtils.extractHeaderValue;
 import static org.folio.spring.tools.config.RetryTemplateConfiguration.DEFAULT_KAFKA_RETRY_TEMPLATE_NAME;
 
+import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
@@ -11,6 +17,7 @@ import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.logging.log4j.message.FormattedMessageFactory;
 import org.folio.entlinks.integration.dto.event.AuthorityDomainEvent;
 import org.folio.entlinks.service.messaging.authority.InstanceAuthorityLinkUpdateService;
+import org.folio.spring.integration.XOkapiHeaders;
 import org.folio.spring.service.SystemUserScopedExecutionService;
 import org.folio.spring.tools.batch.MessageBatchProcessor;
 import org.springframework.kafka.annotation.KafkaListener;
@@ -33,20 +40,29 @@ public class AuthorityEventListener {
   public void handleEvents(List<ConsumerRecord<String, AuthorityDomainEvent>> consumerRecords) {
     log.info("Processing authorities from Kafka events [number of records: {}]", consumerRecords.size());
 
-    var authorityEvents =
-      consumerRecords.stream()
-        .map(consumerRecord -> {
-          var value = consumerRecord.value();
-          value.setId(UUID.fromString(consumerRecord.key()));
-          return value;
-        })
-        .collect(Collectors.groupingBy(AuthorityDomainEvent::getTenant));
-
-    authorityEvents.forEach(this::handleAuthorityEventsForTenant);
+    consumerRecords.stream()
+      .collect(Collectors.groupingBy(consumerRecord ->
+        extractHeaderValue(XOkapiHeaders.USER_ID, consumerRecord.headers())))
+      .forEach(this::handleAuthorityEventsForUser);
   }
 
-  private void handleAuthorityEventsForTenant(String tenant, List<AuthorityDomainEvent> events) {
-    executionService.executeSystemUserScoped(tenant, () -> {
+  private void handleAuthorityEventsForUser(String userId, List<ConsumerRecord<String, AuthorityDomainEvent>> records) {
+    var headers = isBlank(userId)
+      ? Collections.<String, Collection<String>>emptyMap()
+      : Map.<String, Collection<String>>of(XOkapiHeaders.USER_ID, singletonList(userId));
+    records.stream()
+      .map(consumerRecord -> {
+        var value = consumerRecord.value();
+        value.setId(UUID.fromString(consumerRecord.key()));
+        return value;
+      })
+      .collect(Collectors.groupingBy(AuthorityDomainEvent::getTenant))
+      .forEach((tenant, events) -> handleAuthorityEventsForTenant(tenant, headers, events));
+  }
+
+  private void handleAuthorityEventsForTenant(String tenant, Map<String, Collection<String>> headers,
+                                              List<AuthorityDomainEvent> events) {
+    executionService.executeSystemUserScoped(tenant, headers, () -> {
       log.info("Triggering updates for authority records [number of records: {}, tenant: {}]", events.size(), tenant);
       messageBatchProcessor.consumeBatchWithFallback(events, DEFAULT_KAFKA_RETRY_TEMPLATE_NAME,
         instanceAuthorityLinkUpdateService::handleAuthoritiesChanges, this::logFailedEvent);