+++ src/test/java/org/folio/entlinks/integration/kafka/AuthorityEventListenerTest.java
@@ -4,17 +4,22 @@ import static java.util.Collections.singletonList;
 import static org.folio.support.MockingTestUtils.mockBatchFailedHandling;
 import static org.folio.support.MockingTestUtils.mockBatchSuccessHandling;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.Callable;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
+import org.apache.kafka.common.header.internals.RecordHeaders;
 import org.folio.entlinks.domain.dto.AuthorityDto;
 import org.folio.entlinks.domain.dto.Metadata;
 import org.folio.entlinks.integration.dto.event.AuthorityDomainEvent;
 import org.folio.entlinks.service.messaging.authority.InstanceAuthorityLinkUpdateService;
+import org.folio.spring.integration.XOkapiHeaders;
 import org.folio.spring.service.SystemUserScopedExecutionService;
 import org.folio.spring.testing.type.UnitTest;
 import org.folio.spring.tools.batch.MessageBatchProcessor;
@@ -47,8 +52,8 @@ class AuthorityEventListenerTest {
 
   @BeforeEach
   void setUp() {
-    when(executionService.executeSystemUserScoped(any(), any())).thenAnswer(invocation -> {
-      var argument = invocation.getArgument(1, Callable.class);
+    when(executionService.executeSystemUserScoped(any(), any(), any())).thenAnswer(invocation -> {
+      var argument = invocation.getArgument(2, Callable.class);
       return argument.call();
     });
   }
@@ -59,14 +64,19 @@ class AuthorityEventListenerTest {
     var authId = UUID.randomUUID();
     var newRecord = new AuthorityDto().id(authId);
     var oldRecord = new AuthorityDto().id(authId);
+    var userId = UUID.randomUUID().toString();
+    var headers = new RecordHeaders();
+    headers.add(XOkapiHeaders.USER_ID, userId.getBytes());
     var event = TestDataUtils.authorityEvent(type, newRecord, oldRecord);
 
     mockBatchSuccessHandling(messageBatchProcessor);
     when(consumerRecord.key()).thenReturn(authId.toString());
     when(consumerRecord.value()).thenReturn(event);
+    when(consumerRecord.headers()).thenReturn(headers);
 
     listener.handleEvents(singletonList(consumerRecord));
 
+    verify(executionService).executeSystemUserScoped(any(), eq(Map.of(XOkapiHeaders.USER_ID, List.of(userId))), any());
     verify(instanceAuthorityLinkUpdateService).handleAuthoritiesChanges(singletonList(event));
   }